refactor MainActivity.kt repository > viewModel
